+++ test/tests/__init__.py
@@ -65,6 +65,7 @@ AnimationService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["AnimationService"]
 AssetService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["AssetService"]
 ClothesService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["ClothesService"]
 ExportService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["ExportService"]
+FaceService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["FaceService"]
 HumanService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["HumanService"]
 LocationService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["LocationService"]
 LogService = MPFB_CONTEXTUAL_INFORMATION["SERVICES"]["LogService"]
