+++ test/tests/bbb_services/faceservice_test.py
@@ -1,0 +1,10 @@
+import bpy
+from .. import FaceService
+from .. import dynamic_import
+
+# TODO: Write actual tests
+
+def test_faceservice_exists():
+    """FaceService"""
+    assert FaceService is not None, "FaceService can be imported"
+
